refactor: split simulation module into multiple files
